Fixed compiler errors during build of utility_dists library. There are some link-time errors resulting from incomplete explicit template instantiation statements.

+++ packages/utility/distribution/src/Utility_TabularCDFDistribution.hpp
@@ -125,7 +125,7 @@ public:
   IndepQuantity sampleAndRecordTrials( DistributionTraits::Counter& trials ) const override;
 
   //! Return a random sample and bin index from the distribution
-  IndepQuantity sampleAndRecordBinIndex( unsigned& sampled_bin_index ) const override;
+  IndepQuantity sampleAndRecordBinIndex( size_t& sampled_bin_index ) const override;
 
   //! Return a random sample from the distribution at the given CDF value
   IndepQuantity sampleWithRandomNumber( const double random_number ) const override;
@@ -240,7 +240,7 @@ private:
 
   // Return a random sample using the random number and record the bin index
   IndepQuantity sampleImplementation( double random_number,
-                                      unsigned& sampled_bin_index ) const;
+                                      size_t& sampled_bin_index ) const;
 
   // Verify that the values are valid
   template<typename InputIndepQuantity, typename InputDepQuantity>
